+++ src/hkimjp/konpy2/admin.clj
@@ -82,6 +82,8 @@
     [?e :testcode ?testcode]
     [?e :updated ?updated]])
 
+;; FIXME:
+;; prefer `tasks` style. rewrite.
 (defn- div-problems []
   (t/log! :debug "div-problems")
   [:div