+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 # 0.2.16-SNAPSHOT (2025-09-17)
 
+- **BREAKING** attr :test -> :testcode
+
 
 # 0.2.15 (2025-09-17)